+++ examples/floor_division_test.rs
@@ -8,7 +8,9 @@
     let b = 3;
     let q = a / b;
     let r = a % b;
-    let needs_adjustment = r != 0 &&(r<0) ! = (b<0);
+    let r_negative = r<0;
+    let b_negative = b<0;
+    let needs_adjustment = r != 0 && r_negative != b_negative;
     if needs_adjustment {
     q - 1
 }
@@ -26,7 +28,9 @@ else {
     let b = 3;
     let q = a / b;
     let r = a % b;
-    let needs_adjustment = r != 0 &&(r<0) ! = (b<0);
+    let r_negative = r<0;
+    let b_negative = b<0;
+    let needs_adjustment = r != 0 && r_negative != b_negative;
     if needs_adjustment {
     q - 1
 }
@@ -44,7 +48,9 @@ else {
     let b = 3;
     let q = a / b;
     let r = a % b;
-    let needs_adjustment = r != 0 &&(r<0) ! = (b<0);
+    let r_negative = r<0;
+    let b_negative = b<0;
+    let needs_adjustment = r != 0 && r_negative != b_negative;
     if needs_adjustment {
     q - 1
 }
@@ -62,7 +68,9 @@ else {
     let b = 3;
     let q = a / b;
     let r = a % b;
-    let needs_adjustment = r != 0 &&(r<0) ! = (b<0);
+    let r_negative = r<0;
+    let b_negative = b<0;
+    let needs_adjustment = r != 0 && r_negative != b_negative;
     if needs_adjustment {
     q - 1
 }
@@ -80,7 +88,9 @@ else {
     let b = 3;
     let q = a / b;
     let r = a % b;
-    let needs_adjustment = r != 0 &&(r<0) ! = (b<0);
+    let r_negative = r<0;
+    let b_negative = b<0;
+    let needs_adjustment = r != 0 && r_negative != b_negative;
     if needs_adjustment {
     q - 1
 }
@@ -98,7 +108,9 @@ else {
     let b = 3;
     let q = a / b;
     let r = a % b;
-    let needs_adjustment = r != 0 &&(r<0) ! = (b<0);
+    let r_negative = r<0;
+    let b_negative = b<0;
+    let needs_adjustment = r != 0 && r_negative != b_negative;
     if needs_adjustment {
     q - 1
 }